plot works pretty nice deleted one slider precision because the plot couldn't handle it, hope I didn't break anything by it

+++ src/main/java/lazy/AbstractNode.java
@@ -30,7 +30,7 @@ abstract class AbstractNode {
     PVector size = new PVector();
 
     String name;
-    int rowHeightInCells = 1;
+    float rowHeightInCells = 1;
 
     boolean isDragged = false;
     boolean isMouseOverNode = false;
